src/requirements.txt
click == 8.1.7
flask-socketio == 5.3.6
flask[async] == 3.0.2
gevent == 24.2.1
gevent-websocket == 0.10.1
trogon == 0.5.0
textual == 0.52.1
pyopenssl == 24.1.0
PyYAML == 6.0.1
